backend/requirements.txt
fastapi
uvicorn
pymongo
motor
openai>=1.0.0
anthropic>=0.18.0
python-dotenv
